% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeFuns.R
\name{compressedRidge}
\alias{compressedRidge}
\title{Perform compressed (penalized) linear regression}
\usage{
compressedRidge(X, Y, compression = c("xy", "qxqy", "qxy", "linComb",
  "convexComb"), q, lam = NULL, lam.max = NULL, lam.min = 1e-06,
  nlam = 100, s = 3, tol.lam0 = 1e-08, tol.lc = 1e-10)
}
\arguments{
\item{X}{the design matrix \code{(n x p)}}

\item{Y}{the response vector}

\item{compression}{either none ('xy'), full compression ('qxqy'), partial
compression ('qxy'), linear combination ('linComb') or  the convex combination ('convexComb')}

\item{q}{columns in the compression matrix}

\item{lam}{optional values of the tuning parameter. Default is \code{[lam.min, lam.max]} with
\code{n}lam entries equally spaced on the log scale}

\item{lam.max}{defaults to the maximum L1-norm of the covariates divided by 1e-3}

\item{lam.min}{defaults to 1e-6, unless the smallest singular value is larger than \code{tol},
in which case 0 is included}

\item{nlam}{number of lambda values, default is 100, 101 if 0 is included}

\item{s}{\code{1/s} is the probability of non-zero entries in the compression matrix, default
is 3}

\item{tol.lam0}{determines how close to singular the design can be and still try to
use \code{lam=0}. Ignored if \code{lam} is given. Compares with the smallest singular value of
the design.}

\item{tol.lc}{determines how close to singular the matrix of fitted values can be. Compares with
the smallest singular value. Default is then a 50-50 combination of \code{qxy} and \code{qxqy}.}
}
\value{
A list with components of class 'cplr':
 \describe{
 \item{\code{intercept}}{a vector of length \code{nlam} containing the intercept}
   \item{\code{bhat}}{a matrix of size \code{p x nlam} containing the estimated coefficients}
   \item{\code{bhatsc}}{a matrix of size \code{p x nlam} containing the scaled,
   estimated coefficients. For use with the \code{plot} method.}
   \item{\code{fitted}}{a matrix of size \code{n x nlam} containing fitted values}
   \item{\code{residuals}}{a matrix of size \code{n x nlam} containing residuals}
   \item{\code{GCV}}{The generalized cross validation score for model selection.}
   \item{\code{lam}}{The sequence of \code{lam} values used. Either generated or user supplied.}
   \item{\code{df}}{The degrees of freedom of the procedure. If full or partial
    compression, this is the trace of the smoothing matrix. For the other cases,
    the procedure is not a linear smoother, but rather a weighted sum of
    two linear smoothers. In that case, this value is simply the weighted sum
    of the two linear procedures. Note that this likely underestimates the
    true degrees of freedom.}
    \item{\code{train}}{The training error for each value of \code{lam}}
}
}
\description{
Perform compressed (penalized) linear regression
}
\examples{
n = 100
p = 5
q = 50
X = generateX(n, diag(1,p), 'rnorm')
Y = generateY(X, p:1, 'rnorm')
out = compressedRidge(X, Y, 'linComb', q=q, lam.max=1)
}
