% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cplr.R
\docType{package}
\name{cplr}
\alias{cplr}
\alias{cplr-package}
\title{cplr: Compressed and penalized linear regression}
\description{
The cplr package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

