% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ourSimulation.R
\name{createData}
\alias{createData}
\title{Generates data as in the paper}
\usage{
createData(ntrain, ntest, p, covType, rho, bmethod, SNR, noisesd, ...)
}
\arguments{
\item{ntrain}{number of training samples}

\item{ntest}{number of testing samples}

\item{p}{number of predictors}

\item{covType}{for use with \code{\link{generateCov}}}

\item{rho}{for use with \code{\link{generateCov}}}

\item{bmethod}{for use with \code{\link{generateb}}}

\item{SNR}{if not na, sets the standard deviation of the noise to \eqn{\sqrt{b'Cov b} / SNR}}

\item{noisesd}{the standard deviation of the noise if \code{SNR=NA}}

\item{...}{optional arguments passed to \code{generateb}}
}
\value{
A list with components \code{Xtrain}, \code{Ytrain}, \code{Xtest}, \code{Ytest}, \code{bstar}, and \code{sig} giving respectively the training data, testing data, population regression coefficients, and noise standard deviation. The matrix \code{Xtest} is the same as the training data adjusted to the correct size (either by truncating or recycling) so that predictions are at the same design points. Note that this function is simply a wrapper of the various \code{generate*} functions for ease of simulation. If more flexibility is desired, use those directly or similar.
}
\description{
Generates data as in the paper
}
\examples{
createData(2500, 2, 250, 'corrUniform', .8, 'constant', NA, 2500-250-1)
}
