% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divergence.R
\name{divergenceF}
\alias{divergenceF}
\title{Derivative of F}
\usage{
divergenceF(X, Y, QtQ, b_PP, b_FP, lam)
}
\arguments{
\item{X}{the \eqn{n \times p} design matrix}

\item{Y}{the response vector}

\item{QtQ}{the crossproduct of the compression matris}

\item{lam}{the tuning parameter}

\item{b_PC}{the coefficient vector estimated via partial compression}

\item{b_FC}{the coefficient vector estimated via full compression}
}
\value{
A list with two components: \code{divergence} gives the estimated
degrees of freedom, \code{partial_F.partial.y} gives the gradient with respect
to the response \code{Y}.
}
\description{
For approximating the degrees of freedom of the convex combination of full
and partial compression.
}
\details{
\deqn{W = X( t(X)t(Q)QX + lam I)^(-1)t(X)}
\deqn{\hat{Y} = F(y) = W f(y) g(y) h(y)}
\deqn{\Rightarrow F_i(y) = W[i,] f(y) g(y) h(y)}
\deqn{\partial F_i(y) / \partial y_i = W[i,] (df gh + f(dg h + g dh))}
}

