% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{empEstimRisk}
\alias{empEstimRisk}
\title{Calculate the MSE}
\usage{
empEstimRisk(bhat, bstar)
}
\arguments{
\item{bhat}{estimated coefficient vector}

\item{bstar}{true coefficient vector}
}
\value{
The average squared estimation error
\deqn{mean((bhat - bstar)^2).}
}
\description{
Calculate the MSE
}

