% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{empPredRisk}
\alias{empPredRisk}
\title{Calculate the MSPE on new data}
\usage{
empPredRisk(bhat, newX, newY)
}
\arguments{
\item{bhat}{coefficient vector}

\item{newX}{new design matrix}

\item{newY}{new response vector}
}
\value{
The average squared prediction error
\deqn{mean((newY-newX bhat)^2).}
}
\description{
Calculate the MSPE on new data
}
