% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries.R
\name{GCVridge}
\alias{GCVridge}
\title{The generalized cross validation (GCV) score for ridge regression}
\usage{
GCVridge(lam, XX, XY, X, Y)
}
\arguments{
\item{lam}{The tuning parameter}

\item{XX}{The inner product of the design with itself (a pxp matrix)}

\item{XY}{The inner product of the design with the response (a px1 vector)}

\item{X}{The design matrix}

\item{Y}{The response vector}
}
\value{
The generalized cross validation score.
}
\description{
This function is designed to be optimized over using already computed
information. It is not intended to stand alone.
}

