% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskEstimators.R
\name{gcvRisk}
\alias{gcvRisk}
\title{Generalized cross validation risk estimate}
\usage{
gcvRisk(X, Y, bhat, df)
}
\arguments{
\item{X}{the design matrix}

\item{Y}{the response vector}

\item{bhat}{the estimate of the regression coefficients}

\item{df}{the estimated (or exact) degrees of freedom}
}
\value{
The risk estimate.
#@export
}
\description{
Generalized cross validation risk estimate
}
\details{
The calculation is given by
 \deqn{\sum (X \hat{beta} - Y)^2 / (1 - df/n)^2}
}
