% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateModel.R
\name{generateb}
\alias{generateb}
\title{Generates a vector of population regression coefficients}
\usage{
generateb(p, method = "randN", b1 = NULL, s = 10, spar = 10, ...)
}
\arguments{
\item{p}{the number of coefficients}

\item{method}{one of eight named methods of generation, see Details below}

\item{b1}{for \code{method='expdecay'}, see Details below}

\item{s}{for \code{method='sparse'}, see Details below}

\item{spar}{the l2 norm of the resulting coefficient vector, see Details below}

\item{...}{optional arguments to random number generators, see Details below}
}
\value{
a vector of length p
}
\description{
Generates a vector of population regression coefficients
}
\details{
This function generates a vector of coefficients using one of eight methods.
\describe{
 \item{expdecay}{The result is \code{b1^(1:p)}}
 \item{lindecay}{p:1/p}
 \item{sparse}{Produces \code{s} coefficents drawn independently from the Laplace distribution, followed by \code{p-s} zeros}
 \item{randU}{uniform random coefficents, takes arguments in \code{...}}
 \item{randN}{Random Normal coefficients, takes arguments in \code{...}}
 \item{constNorm}{all coefficients are equal and the vector has l2 norm \code{spar}}
 \item{constLinDecay}{Same as \code{lindecay} except that the resulting vector has norm \code{spar}}
 \item{constant}{produces a p-vector of 1's}
 }
}
\examples{
generateb(10, 'constNorm')
generateb(10, 'randN', mean=1, sd=.1)
generateb(10, 'expdecay', b1=.9)
generateb(10, 'expdecay', b1=1) # same as 'constant'
generateb(10, 'constant')
}

