% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateModel.R
\name{generateCov}
\alias{generateCov}
\title{Generates the square root of the covariance matrix}
\usage{
generateCov(p, type = c("diag", "corrUniform", "corrDecay"), rho = 0.8)
}
\arguments{
\item{p}{the number of predictors}

\item{type}{\code{'diag'} (the default) gives the identity matrix, \code{'corrUniform'} gives a matrix with 1 on the diagonal and \code{rho} off the diagonal, and \code{'corrDecay'} gives the AR covariance matrix (1 on the diagonal, \eqn{\rho^|i-j|} off the diagonal)}

\item{rho}{the correlation, ignored for \code{type='diag'}}
}
\value{
Returns the square root of the covariance
}
\description{
Generates the square root of the covariance matrix
}
\examples{
generateCov(10,'diag')
generateCov(10, 'corrDecay')
generateCov(10, 'corrDecay',.9)
}

