% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsFuncs.R
\name{generateQ}
\alias{generateQ}
\title{Generate a preconditioner}
\usage{
generateQ(q, n, rand = "rsparseBern", normalize = sqrt(q/3), ...)
}
\arguments{
\item{q}{number of rows}

\item{n}{number of columns}

\item{rand}{quoted name of a random number generator, defaults to \code{'rsparseBern'}}

\item{normalize}{optional scaler to normalize the matrix, defaults
to \eqn{\sqrt{q/3}}, which makes \eqn{E[Q'Q]=I} for sparse Bernoullis
with \code{s=3}, see \code{\link{rsparseBern}}}

\item{...}{optional additional arguments passed to the random number generator}
}
\value{
A qxn matrix of random numbers
}
\description{
Generate a preconditioner
}
\examples{
generateQ(5, 10)
}

