% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateModel.R
\name{generateX}
\alias{generateX}
\title{Generates the design matrix}
\usage{
generateX(n, covMatSqrt, rand = "rnorm", m = NULL, ...)
}
\arguments{
\item{n}{number of observations}

\item{covMatSqrt}{square root of the design covariance, possibly produced by \code{\link{generateCov}}}

\item{rand}{character string giving the name of a function generating random numbers}

\item{m}{optional mean vector for the design matrix}

\item{...}{optional named arguments passed to the random number generating function}
}
\value{
Returns a \code{n x ncol(covMatSqrt)} matrix.
}
\description{
Generates the design matrix
}
\examples{
generateX(100, diag(1,10), 'rsparseBern')
}
