% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes}
\alias{genes}
\alias{b1}
\alias{b2}
\alias{b3}
\alias{g1}
\alias{g2}
\alias{w1}
\alias{w2}
\alias{w3}
\title{Genetics datasets}
\format{Each of the eight datasets contains 40 columns. Columns 2 through 40 contain the 20 nucleotides to the left and 19 nucleotides to the right of a particular nucleotide. The first column, \code{count}, contains the number of times those sequences were found in different RNA sequences.

The datasets beginning with \code{b} are human and originated with the Burge group \url{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC2593745/}.

The datasets beginning with \code{g} are from mice and due to the Grimmond group \url{http://www.nature.com/nmeth/journal/v5/n7/full/nmeth.1223.html}.

The datasets beginning with \code{w} are also from mice and due to the Wold group \url{http://www.nature.com/nmeth/journal/v5/n7/full/nmeth.1226.html}.}
\source{
These data were downloaded from \href{http://www3.nd.edu/~jli9/mseq/data_top100.zip}{Prof. Jun Li} and then processed using his package \href{https://cran.r-project.org/web/packages/mseq/index.html}{mseq}.
}
\usage{
data(b1)

b1

b2

b3

g1

g2

w1

w2

w3
}
\description{
Eight different genetics datasets \code{b1, b2, b3, g1, g2, w1, w2, w3} are used in this paper
}
\keyword{datasets}
