% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries.R
\name{logScaleSeq}
\alias{logScaleSeq}
\title{Create a sequence of numbers equally spaced on the log scale}
\usage{
logScaleSeq(from = 1, to = 1, len = 1, integers = TRUE)
}
\arguments{
\item{from}{the left endpoint}

\item{to}{the right endpoint}

\item{len}{the length of the resulting sequence, a non-negative number}

\item{integers}{if TRUE, rounds the result is rounded to the nearest integer}
}
\value{
a vector of length \code{len}
}
\description{
Create a sequence of numbers equally spaced on the log scale
}
\examples{
logScaleSeq(1,1000,20)
logScaleSeq(1,1000,20, FALSE)
}

