% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neuro}
\alias{neuro}
\title{T1 weighted MR image of a brain}
\format{This data set is a 181x217x181 array of intensity values on the range of 0 to 255.}
\source{
These data were downloaded from 
\href{http://brainweb.bic.mni.mcgill.ca/selection_normal.html}{BrainWeb} 
using the T1 modality, 1mm slice thickness, 0\% noise, and 20\% intensity 
non-uniformity. We selected the "raw byte" format. Finally, we created this 
array using

\code{neuro = readBin("t1_icbm_normal_1mm_pn0_rf20.rawb", 
                      "integer", 200^3, 1, signed=FALSE)}
\code{dim(neuro) <- c(181,217,181)}
}
\usage{
data(neuro)

neuro
}
\description{
A single simulated image of a brain in a 3D array
}
\keyword{datasets}
