% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{olsPredAlgo}
\alias{olsPredAlgo}
\title{Calculates degrees of freedom for OLS}
\usage{
olsPredAlgo(job, static, dynamic, ...)
}
\arguments{
\item{job}{see \code{\link{BatchExperiments}}}

\item{static}{see \code{\link{BatchExperiments}}}

\item{dynamic}{see \code{\link{BatchExperiments}}}

\item{...}{see \code{\link{BatchExperiments}}}
}
\value{
A list with components:
\describe{
 \item{\code{train}}{the training error}
 \item{\code{test}}{the testing error}
 \item{\code{df}}{the degrees of freedom using the
 smoothing matrix}
 \item{\code{steinR}}{the SURE estimate of the risk using \code{df}}
 \item{\code{gcvR}}{the GCV estimate of the risk using \code{df}}
 }
}
\description{
Calculates degrees of freedom for OLS
}
