% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{plsAlgorithm}
\alias{plsAlgorithm}
\title{Calculates prediction error for PLS}
\usage{
plsAlgorithm(job, static, dynamic, ...)
}
\arguments{
\item{job}{see \code{\link{BatchExperiments}}}

\item{static}{see \code{\link{BatchExperiments}}}

\item{dynamic}{see \code{\link{BatchExperiments}}}

\item{...}{see \code{\link{BatchExperiments}}}
}
\value{
A list with components \code{risk}, \code{bias}, \code{var}, \code{noiseVar}, and \code{ptm}. These give, respectively, the prediction risk, the bias, the variance, the noise variance, and the computation time.
}
\description{
This function is used by \code{\link{BatchExperiments}} to run the simulations.
See that package for documentation.
}
\seealso{
empPredRisk, predBias, predVar
}

