% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsFuncs.R
\name{plsFast}
\alias{plsFast}
\title{Fast version of pls}
\usage{
plsFast(X, Y, compression, q, lam, s = 3)
}
\arguments{
\item{X}{the design matrix (n x p)}

\item{Y}{the response vector (n x 1)}

\item{compression}{the compression type (\code{'qxqy'}, \code{'qxy'}, \code{'linComb'}, \code{'convexComb'})}

\item{q}{the compression parameter}

\item{lam}{the ridge penalty}

\item{s}{optional sparsity parameter for the compression matrix, defaults
to 3}
}
\value{
A list with components \code{bhat} (estimated coefficients) and
 \code{df} (the trace of the hat matrix, possibly weighted)
#@export
}
\description{
This version uses a C function to perform the compression without ever
generating the Q matrix. There are fewer options that with \code{\link{pls}}
for this reason. The divergence-based risk estimator is unavailable.
An estimate of the degrees-of-freedom is returned.
}
\seealso{
pls
}
