% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms_deprecated.R
\name{plsRiskEstimAlgo}
\alias{plsRiskEstimAlgo}
\title{Calculates degrees of freedom for compressed methods}
\usage{
plsRiskEstimAlgo(job, static, dynamic, ...)
}
\arguments{
\item{job}{see \code{BatchExperiments}}

\item{static}{see \code{BatchExperiments}}

\item{dynamic}{see \code{BatchExperiments}}

\item{...}{see \code{BatchExperiments}}
}
\value{
A list with components:
\describe{
 \item{\code{df}}{the (exact or approximate) degrees of freedom using the
 smoothing matrix}
 \item{\code{steinR}}{the SURE estimate of the risk using \code{df}}
 \item{\code{gcvR}}{the GCV estimate of the risk using \code{df}}
 \item{\code{divdf}}{the divergence-based degrees-of-freedom estimator for
 the combined methods}
 \item{\code{steindivR}}{the SURE estimate of the risk using \code{divdf}}
 \item{\code{gcvdivR}}{the GCV estimate of the risk using \code{divdf}}
 \item{\code{train}}{the training error}
 \item{\code{test}}{the testing error}
 }

#@export
}
\description{
Calculates degrees of freedom for compressed methods
}
