% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precondls.R
\docType{package}
\name{precondls}
\alias{precondls}
\alias{precondls-package}
\title{precondls: Preconditioned (penalized) least-squares regression}
\description{
The precondls package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

