% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{predBias}
\alias{predBias}
\title{Calculate the average squared bias (per prediction)}
\usage{
predBias(bstar, newX, XstuffTrain, bhat = NULL)
}
\arguments{
\item{bstar}{true coefficient vector}

\item{newX}{new design matrix}

\item{XstuffTrain}{the portion of the estimated coefficient vector that depends
on the design matrix from the training set or \code{NULL}.}

\item{bhat}{estimated coefficient vector}
}
\value{
The average squared difference between the predictions and their
expectation
}
\description{
Calculate the average squared bias (per prediction)
}
\details{
If \code{XstuffTrain} is given, the linear smoother gives the
 \eqn{\E[\hat{Y}]}. Otherwise (NULL), evaluate the predictions on the
 test data to get monte carlo bias
}

