% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{predVar}
\alias{predVar}
\title{Calculate the average variance (per prediction)}
\usage{
predVar(sigma, newX, XstuffTrain, bhat = NULL)
}
\arguments{
\item{sigma}{the standard deviation of the noise}

\item{newX}{new design matrix}

\item{XstuffTrain}{the portion of the estimated coefficient vector that depends
on the design matrix from the training set or \code{NULL}.}

\item{bhat}{estimated coefficient vector or \code{NULL}}
}
\value{
The average variance per prediction
}
\description{
Calculate the average variance (per prediction)
}
\details{
If \code{XstuffTrain} is given, the linear smoother gives the
 variance exactly as the trace of the squared "hat" matrix (not necessarily
 idempotent). Otherwise (NULL), evaluate the predictions on the
 test data to get monte carlo variance
}

