% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{ridgeEstimAlgo}
\alias{ridgeEstimAlgo}
\title{Calculates estimation error for ridge regression (optimal lambda)}
\usage{
ridgeEstimAlgo(job, static, dynamic, ...)
}
\arguments{
\item{job}{see \code{\link{BatchExperiments}}}

\item{static}{see \code{\link{BatchExperiments}}}

\item{dynamic}{see \code{\link{BatchExperiments}}}

\item{...}{see \code{\link{BatchExperiments}}}
}
\value{
A list with components \code{estimRisk} and \code{ptm}. The first gives the estimation risk of the particular method while the second calculates how long the solution took to compute.
}
\description{
This function is used by \code{\link{BatchExperiments}} to run the simulations.
See that package for documentation.
}
\seealso{
empEstimRisk
}

