% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{ridgePredAlgo}
\alias{ridgePredAlgo}
\title{Calculates degrees of freedom for ridge regression}
\usage{
ridgePredAlgo(job, static, dynamic, ...)
}
\arguments{
\item{job}{see \code{BatchExperiments}}

\item{static}{see \code{BatchExperiments}}

\item{dynamic}{see \code{BatchExperiments}}

\item{...}{see \code{BatchExperiments}}
}
\value{
A list with components:
\describe{
 \item{\code{test}}{the testing error}
 \item{\code{train}}{the training error}
 \item{\code{df}}{the degrees of freedom using the
 smoothing matrix}
 \item{\code{gcvR}}{the GCV estimate of the risk using \code{df}}
 \item{\code{steinR}}{the SURE estimate of the risk using \code{df}}
 \item{\code{lamStar}}{the lambda value used, chosen via GCV}
 }
}
\description{
Calculates degrees of freedom for ridge regression
}
