% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries.R
\name{ridgeRegression}
\alias{ridgeRegression}
\title{No intercept ridge regression, optimal lambda}
\usage{
ridgeRegression(X, Y, lam.upper = min(dim(X)))
}
\arguments{
\item{X}{design matrix}

\item{Y}{response vector}

\item{lamLimits}{a vector with the smallest and largest lambda values to try}
}
\value{
A list containing th vector of coefficients at the GCV minimizing value of lambda (\code{bhat}) and the best value of lambda (\code{lamStar}).
}
\description{
Computes the ridge regression estimate, optimizing over the tuning parameter
lambda using generalized cross validation (GCV)
}
