% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries.R
\name{ridgeRegressionLam}
\alias{ridgeRegressionLam}
\title{No intercept ridge regression}
\usage{
ridgeRegressionLam(X, Y, lam)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response vector}

\item{lam}{a value of the penalty parameter}
}
\value{
A vector of coefficients
}
\description{
Computes the ridge regression estimate for a fixed value of lambda.
}
