% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateModel.R
\name{rlaplace}
\alias{rlaplace}
\title{Generates standard Lapace random variates}
\usage{
rlaplace(n)
}
\arguments{
\item{n}{the number of desired random numbers}
}
\value{
an n-vector of independent Laplace random variates
}
\description{
Generates standard Lapace random variates
}

