% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsFuncs.R
\name{rsparseBern}
\alias{rsparseBern}
\title{Generate sparse Bernoulli random variables}
\usage{
rsparseBern(n, s = 3)
}
\arguments{
\item{n}{the number of random variates}

\item{s}{calibrates the probability of non-zeros. Default value is 3. Should be greater than one or results in unexpected behavior.}
}
\value{
Vector of random sparse Bernoulli draws
}
\description{
The sparse Bernoulli distribution puts mass on the set \eqn{\{-1, 0, 1\}} with probabilities \eqn{\{1/2s, 1-1/s, 1/2s\}}.
}
\examples{
rsparseBern(10)
rsparseBern(10, 10)
}

