% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskEstimators.R
\name{steinRisk}
\alias{steinRisk}
\title{Computes Stein's Risk Estimate}
\usage{
steinRisk(X, Y, bhat, df)
}
\arguments{
\item{X}{the design matrix}

\item{Y}{the response vector}

\item{bhat}{the estimate of the regression coefficients}

\item{df}{the estimated (or exact) degrees of freedom}
}
\value{
The risk estimate. Note that this is not necessarily unbiased as the variance is unknown
 and the estimator used here is not necessarily unbiased.
#@export
}
\description{
Computes Stein's Risk Estimate
}
